ChatWithPDf/requirements.txt
langchain
langchain-core
langchain-community
langchain-text-splitters
pypdf



langchain-huggingface
transformers
sentence-transformers
huggingface-hub

python-dotenv

streamlit

faiss-cpu

PyPDF2==3.0.1